Move control of run co-routine from task runner
Change the task runner to return a generator to be passed to the
task run() method.  The task role now is to send() status to the
runner which will report back if the task has been canceled with
StopIteration.  This allows the backend to wait for responses
from the child urpmi process and signal the runner when
appropriate, sparing loops in the main loop.

Add a decorator for run() method of tasks that won't use the
backend so they can signal the runner accordingly.

Some other minor changes related:

- Use self.canceled to signal task cancellation instead of state,
  so the task state won't change to STATE_CANCELING before the
  run() method receives StopIteration.

- Add STATE_RUNNING state to guarantee that the task won't be
  STATE_QUEUED after it has been removed from the queue.

- Typo in _remove_and_cleanup method call.

- Add on_error() callback to handle task errors.