Replace URPM::Package by the response arg
Do not save URPM::Package in memory to extract the response arg later, instead
produce the response arg and save it in the pkg_map.  To simplify this create a
standard way to make package reference in backend responses, a tuple of (na,
evrd).  This force the need of update in the mdvpkg handlers.